Refactor FeatureStream into wildfly-version.
