+++ controller/src/main/java/org/jboss/as/controller/FeatureStreamAware.java
@@ -4,6 +4,8 @@
  */
 package org.jboss.as.controller;
 
+import org.jboss.as.version.FeatureStream;
+
 /**
  * Implemented by objects that should be aware of the configured feature stream.
  * @author Paul Ferraro
