+++ controller/src/main/java/org/jboss/as/controller/xml/SimpleVersionedNamespace.java
@@ -4,7 +4,7 @@
  */
 package org.jboss.as.controller.xml;
 
-import org.jboss.as.controller.FeatureStream;
+import org.jboss.as.version.FeatureStream;
 import org.jboss.staxmapper.Versioned;
 
 /**
