+++ controller/src/test/java/org/jboss/as/controller/xml/VersionedNamespaceTestCase.java
@@ -6,8 +6,8 @@ package org.jboss.as.controller.xml;
 
 import java.util.List;
 
-import org.jboss.as.controller.FeatureStream;
 import org.jboss.as.controller.SubsystemSchema;
+import org.jboss.as.version.FeatureStream;
 import org.jboss.staxmapper.IntVersion;
 import org.junit.Assert;
 import org.junit.Test;
