+++ core-feature-pack/common/src/main/resources/modules/system/layers/base/org/jboss/as/controller/main/module.xml
@@ -25,6 +25,7 @@
         <module name="org.jboss.as.controller-client" export="true"/>
         <module name="org.jboss.as.core-security"/>
         <module name="org.jboss.as.protocol"/>
+        <module name="org.jboss.as.version"/>
         <module name="org.jboss.dmr" export="true"/>
         <module name="org.jboss.logging"/>
         <module name="org.jboss.logmanager"/>
